+++ reikna/cluda/array_helpers.py
@@ -10,7 +10,7 @@ import reikna.cluda.dtypes as dtypes
 import reikna.transformations as transformations
 import reikna.cluda.functions as functions
 from reikna.algorithms import PureParallel
-from reikna.core import Type
+from reikna.core import Type, Parameter, Annotation, Computation
 
 
 def normalize_value(thr, gpu_array_type, val):
@@ -121,3 +121,79 @@ def concatenate(arrays, axis=0, out=None):
         offset += array.shape[axis]
 
     return out
+
+
+def roll_computation(array, axis):
+    return PureParallel(
+        [
+            Parameter('output', Annotation(array, 'o')),
+            Parameter('input', Annotation(array, 'i')),
+            Parameter('shift', Annotation(Type(numpy.int32)))],
+        """
+        <%
+            shape = input.shape
+        %>
+        %for i in range(len(shape)):
+            VSIZE_T output_${idxs[i]} =
+                %if i == axis:
+                ${shift} == 0 ?
+                    ${idxs[i]} :
+                    ## Since ``shift`` can be negative, and its absolute value greater than
+                    ## ``shape[i]``, a double modulo division is necessary
+                    ## (the ``%`` operator preserves the sign of the dividend in C).
+                    (${idxs[i]} + (${shape[i]} + ${shift} % ${shape[i]})) % ${shape[i]};
+                %else:
+                ${idxs[i]};
+                %endif
+        %endfor
+        ${output.store_idx}(
+            ${", ".join("output_" + name for name in idxs)},
+            ${input.load_idx}(${", ".join(idxs)}));
+        """,
+        guiding_array='input',
+        render_kwds=dict(axis=axis))
+
+
+class RollInplace(Computation):
+
+    def __init__(self, array, axis):
+        self._axis = axis
+        Computation.__init__(self, [
+            Parameter('array', Annotation(array, 'io')),
+            Parameter('shift', Annotation(Type(numpy.int32)))])
+
+    def _build_plan(self, plan_factory, device_params, array, shift):
+        plan = plan_factory()
+
+        temp = plan.temp_array_like(array)
+        plan.computation_call(roll_computation(array, self._axis), temp, array, shift)
+
+        tr = transformations.copy(temp, out_arr_t=array)
+        copy_comp = PureParallel.from_trf(tr, guiding_array=tr.output)
+        plan.computation_call(copy_comp, array, temp)
+
+        return plan
+
+
+def roll(array, shift, axis=-1):
+    """
+    Cyclically shifts elements of ``array`` by ``shift`` positions to the right along ``axis``.
+    ``shift`` can be negative (in which case the elements are shifted to the left).
+    Elements that are shifted beyond the last position are re-introduced at the first
+    (and vice versa).
+
+    Works equivalently to ``numpy.roll`` (except ``axis=None`` is not supported).
+    """
+    temp = array.thread.array(array.shape, array.dtype)
+    axis = axis % len(array.shape)
+    comp = array.thread.get_cached_computation(
+        roll_computation, Type.from_value(array), axis)
+    comp(temp, array, shift)
+    return temp
+
+
+def roll_method(array, shift, axis=-1):
+    axis = axis % len(array.shape)
+    comp = array.thread.get_cached_computation(
+        RollInplace, Type.from_value(array), axis)
+    comp(array, shift)