+++ doc/source/history.rst
@@ -24,6 +24,8 @@ Release history
 
 * ADDED: :py:func:`~reikna.concatenate` for concatenating GPU arrays.
 
+* ADDED: :py:func:`~reikna.roll` for GPU arrays and the inplace version :py:meth:`~reikna.cluda.api.Array.roll`.
+
 * FIXED: updated the CUDA backend for the change ``async`` -> ``async_`` in the new versions of PyCUDA. Bumped PyOpenCL and PyCUDA versions to 2018.1.1.
 
 * FIXED: an error in the conversion of `numpy.int64` to ctype for Windows.